+++ src/brushes/pattern_brush.class.js
@@ -47,8 +47,8 @@ fabric.PatternBrush = fabric.util.createClass(fabric.PencilBrush, /** @lends fab
   /**
    * Creates path
    */
-  createPath: function(pathData) {
-    var path = this.callSuper('createPath', pathData),
+  createPath: async function(pathData) {
+    var path = await this.callSuper('createPath', pathData),
         topLeft = path._getLeftTopCoords().scalarAdd(path.strokeWidth / 2);
 
     path.stroke = new fabric.Pattern({